Battle hardening tests
- some tests are failing intermittently; been sprinkling cleanup statements where nescessary
- removed old baw-audio-tools and baw-workers folders now that the port is pretty much complete
- added a feature to disable sending user creation emails. This is only used for testing; creation.rb creates users 1000s of times  over the course of a full test suite
- reverted stdout/stderr change: both streams are directed to log files when testting, however I print out a warning now so it is at least obvious!

+++ spec/lib/gems/baw_workers/audio_check/action_spec.rb
@@ -146,7 +146,7 @@ describe BawWorkers::AudioCheck::Action do
           }
 
         audio_original.possible_paths(media_request_params).each do |file|
-          File.delete file
+          File.delete file if File.exist? file
         end
 
         expect {
